% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{dummy_function1}
\alias{dummy_function1}
\title{Dummy Function 1
This function returns a placeholder message.}
\usage{
dummy_function1()
}
\value{
A character string.
}
\description{
Dummy Function 1
This function returns a placeholder message.
}
