% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{dummy_function2}
\alias{dummy_function2}
\title{Dummy Function 2
This function returns a placeholder message.}
\usage{
dummy_function2()
}
\value{
A character string.
}
\description{
Dummy Function 2
This function returns a placeholder message.
}
