% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{dummy_function3}
\alias{dummy_function3}
\title{Dummy Function 3
This function returns a placeholder message.}
\usage{
dummy_function3()
}
\value{
A character string.
}
\description{
Dummy Function 3
This function returns a placeholder message.
}
