% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{dummy_function4}
\alias{dummy_function4}
\title{Dummy Function 4
This function returns a placeholder message.}
\usage{
dummy_function4()
}
\value{
A character string.
}
\description{
Dummy Function 4
This function returns a placeholder message.
}
