% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{dummy_function5}
\alias{dummy_function5}
\title{Dummy Function 5
This function returns a placeholder message.}
\usage{
dummy_function5()
}
\value{
A character string.
}
\description{
Dummy Function 5
This function returns a placeholder message.
}
